# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "API documentation for products, cart and orders endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current application server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create order from current cart",
                "operationId": "storeOrder",
                "requestBody": {
                    "description": "Delivery address for the order",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOrderRequest"
                            },
                            "example": {
                                "region": "Moscow",
                                "city": "Moscow",
                                "street": "Tverskaya",
                                "house": "10",
                                "entrance": "2",
                                "apartment": "45",
                                "postcode": "125009"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                },
                                "example": {
                                    "id": 9,
                                    "user_id": 4,
                                    "address_id": 7,
                                    "status": "created",
                                    "total_price": 998,
                                    "created_at": "2026-04-03T12:15:00.000000Z",
                                    "updated_at": "2026-04-03T12:15:00.000000Z",
                                    "items": [
                                        {
                                            "id": 11,
                                            "order_id": 9,
                                            "product_id": 1,
                                            "quantity": 2,
                                            "price": 499,
                                            "created_at": "2026-04-03T12:15:00.000000Z",
                                            "updated_at": "2026-04-03T12:15:00.000000Z",
                                            "product": {
                                                "id": 1,
                                                "name": "Margherita",
                                                "description": "Classic pizza with mozzarella and tomato sauce.",
                                                "price": 499,
                                                "weight": 450,
                                                "category": "pizza",
                                                "created_at": "2026-04-03 12:00:00",
                                                "updated_at": "2026-04-03 12:00:00"
                                            }
                                        }
                                    ],
                                    "address": {
                                        "id": 7,
                                        "user_id": 4,
                                        "region": "Moscow",
                                        "city": "Moscow",
                                        "street": "Tverskaya",
                                        "house": "10",
                                        "entrance": "2",
                                        "apartment": "45",
                                        "postcode": "125009",
                                        "created_at": "2026-04-03T12:15:00.000000Z",
                                        "updated_at": "2026-04-03T12:15:00.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User role is not allowed to create orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation or business rule error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ValidationErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/BusinessErrorResponse"
                                        }
                                    ]
                                },
                                "examples": {
                                    "validation": {
                                        "summary": "Missing required fields",
                                        "value": {
                                            "message": "The given data was invalid.",
                                            "errors": {
                                                "postcode": [
                                                    "The postcode field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "emptyCart": {
                                        "summary": "Cart is empty",
                                        "value": {
                                            "error": "В корзине нет товаров"
                                        }
                                    },
                                    "limitsExceeded": {
                                        "summary": "Cart limits exceeded",
                                        "value": {
                                            "error": "Превышен лимит товаров"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProductResource": {
                "required": [
                    "id",
                    "name",
                    "price",
                    "category",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Margherita"
                    },
                    "description": {
                        "type": "string",
                        "example": "Classic pizza with mozzarella and tomato sauce.",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 499
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 450,
                        "nullable": true
                    },
                    "category": {
                        "type": "string",
                        "example": "pizza"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-03 12:00:00",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-03 12:00:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreOrderRequest": {
                "required": [
                    "region",
                    "city",
                    "street",
                    "house",
                    "postcode"
                ],
                "properties": {
                    "region": {
                        "type": "string",
                        "example": "Moscow"
                    },
                    "city": {
                        "type": "string",
                        "example": "Moscow"
                    },
                    "street": {
                        "type": "string",
                        "example": "Tverskaya"
                    },
                    "house": {
                        "type": "string",
                        "example": "10"
                    },
                    "entrance": {
                        "type": "string",
                        "example": "2",
                        "nullable": true
                    },
                    "apartment": {
                        "type": "string",
                        "example": "45",
                        "nullable": true
                    },
                    "postcode": {
                        "type": "string",
                        "example": "125009"
                    }
                },
                "type": "object"
            },
            "AddressResponse": {
                "required": [
                    "id",
                    "user_id",
                    "region",
                    "city",
                    "street",
                    "house",
                    "postcode",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "region": {
                        "type": "string",
                        "example": "Moscow"
                    },
                    "city": {
                        "type": "string",
                        "example": "Moscow"
                    },
                    "street": {
                        "type": "string",
                        "example": "Tverskaya"
                    },
                    "house": {
                        "type": "string",
                        "example": "10"
                    },
                    "entrance": {
                        "type": "string",
                        "example": "2",
                        "nullable": true
                    },
                    "apartment": {
                        "type": "string",
                        "example": "45",
                        "nullable": true
                    },
                    "postcode": {
                        "type": "string",
                        "example": "125009"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-03T12:15:00.000000Z",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-03T12:15:00.000000Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderItemResponse": {
                "required": [
                    "id",
                    "order_id",
                    "product_id",
                    "quantity",
                    "price",
                    "created_at",
                    "updated_at",
                    "product"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 11
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 9
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 499
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-03T12:15:00.000000Z",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-03T12:15:00.000000Z",
                        "nullable": true
                    },
                    "product": {
                        "$ref": "#/components/schemas/ProductResource"
                    }
                },
                "type": "object"
            },
            "OrderResponse": {
                "required": [
                    "id",
                    "user_id",
                    "address_id",
                    "status",
                    "total_price",
                    "created_at",
                    "updated_at",
                    "items",
                    "address"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 9
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "address_id": {
                        "type": "integer",
                        "example": 7
                    },
                    "status": {
                        "type": "string",
                        "example": "created",
                        "enum": [
                            "created",
                            "paid",
                            "in_progress",
                            "delivering",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 998
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-03T12:15:00.000000Z",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-03T12:15:00.000000Z",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemResponse"
                        }
                    },
                    "address": {
                        "$ref": "#/components/schemas/AddressResponse"
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ForbiddenResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "quantity": [
                                "The quantity field must be at least 1."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "BusinessErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "В корзине нет товаров"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token passed in the Authorization header as Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}